membuat order langsung endpoint

+++ src/main/java/com/programmerpagi/toko_online/dto/ProductRequestDTO.java
@@ -4,6 +4,8 @@ import lombok.Getter;
 import lombok.Setter;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 
@@ -12,7 +14,7 @@ public class ProductRequestDTO {
     private Long id;
     private String nama;
     private String kategori;
-    private Long harga;
+    private BigDecimal harga;
     private Integer stok;
     private String deskripsi;
     private MultipartFile image;
