+++ src/main/java/com/programmerpagi/toko_online/dto/ProductResponseDTO.java
@@ -3,6 +3,8 @@ package com.programmerpagi.toko_online.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 
@@ -11,7 +13,7 @@ public class ProductResponseDTO {
     private Long id;
     private String nama;
     private String kategori;
-    private Long harga;
+    private BigDecimal harga;
     private Integer stok;
     private String deskripsi;
     private String image;
